+++ src/main.ts
@@ -16,7 +16,50 @@ export default class PomodoroPlugin extends Plugin {
 
 		this.addSettingTab(new PomodoroSettingTab(this.app, this));
 
-		// Add command to toggle status bar visibility
+		// Add commands for keyboard shortcuts
+		this.addCommand({
+			id: 'toggle-timer',
+			name: 'Start/Pause timer',
+			callback: () => {
+				if (this.timer) {
+					if (this.timer.running) {
+						this.timer.pauseTimer();
+					} else {
+						this.timer.startTimer();
+					}
+				}
+			}
+		});
+
+		this.addCommand({
+			id: 'reset-timer',
+			name: 'Reset current timer',
+			callback: () => {
+				if (this.timer && !this.timer.running) {
+					this.timer.resetTimer();
+				}
+			}
+		});
+
+		this.addCommand({
+			id: 'cycle-timer',
+			name: 'Cycle to next timer duration',
+			callback: () => {
+				if (this.timer && !this.timer.running) {
+					this.timer.cycleDuration();
+				}
+			}
+		});
+
+		this.addCommand({
+			id: 'toggle-icon-visibility',
+			name: 'Toggle timer icon visibility',
+			callback: () => {
+				this.settings.showIcon = !this.settings.showIcon;
+				this.saveSettings();
+			}
+		});
+
 		this.addCommand({
 			id: 'toggle-status-bar',
 			name: 'Toggle status bar visibility',